Typos. Rename energy balance model classes for consistency.
