+++ src/HealthGPS/CMakeLists.txt
@@ -24,10 +24,10 @@ target_sources(HealthGPS
 		"disease.h"
 		"disease_table.cpp"
 		"disease_table.h"
+		"energy_balance_hierarchical_model.cpp"
+		"energy_balance_hierarchical_model.h"
 		"energy_balance_model.cpp"
 		"energy_balance_model.h"
-		"energy_balance_model_new.cpp"
-		"energy_balance_model_new.h"
 		"error_message.cpp"
 		"error_message.h"
 		"event_bus.cpp"
