+++ src/HealthGPS/energy_balance_model.h
@@ -5,16 +5,16 @@
 
 namespace hgps {
 
-	/// @brief Implements the dynamic hierarchical linear model (energy balance) type
+	/// @brief Implements the energy balance model type
 	/// 
 	/// @details The dynamic model is used to advance the virtual population over time.
-	class EnergyBalanceHierarchicalModel final : public HierarchicalLinearModel {
+	class EnergyBalanceModel final : public HierarchicalLinearModel {
 	public:
-		EnergyBalanceHierarchicalModel() = delete;
+		EnergyBalanceModel() = delete;
 
-		/// @brief Initialises a new instance of the EnergyBalanceHierarchicalModel class
+		/// @brief Initialises a new instance of the EnergyBalanceModel class
 		/// @param definition The model definition instance
-		EnergyBalanceHierarchicalModel(LiteHierarchicalModelDefinition& definition);
+		EnergyBalanceModel(EnergyBalanceModelDefinition& definition);
 
 		HierarchicalModelType type() const noexcept override;
 
@@ -27,17 +27,7 @@ namespace hgps {
 		void update_risk_factors(RuntimeContext& context) override;
 
 	private:
-		std::reference_wrapper<LiteHierarchicalModelDefinition> definition_;
+		std::reference_wrapper<EnergyBalanceModelDefinition> definition_;
 		std::string name_{ "Dynamic" };
-
-		void update_risk_factors_exposure(RuntimeContext& context, Person& entity,
-			const std::map<core::Identifier, double>& current_risk_factors,
-			const std::map<core::Identifier, FactorDynamicEquation>& equations);
-
-		std::map<core::Identifier, double> get_current_risk_factors(
-			const HierarchicalMapping& mapping, Person& entity, int time_year) const;
-
-		double sample_normal_with_boundary(Random& random,
-			double mean, double standard_deviation, double boundary) const;
 	};
 }
