+++ src/HealthGPS/hierarchical_model.h
@@ -1,5 +1,5 @@
 #pragma once
 
 #include "hierarchical_model_static.h"
+#include "energy_balance_hierarchical_model.h"
 #include "energy_balance_model.h"
-#include "energy_balance_model_new.h"
